requirements.txt
aiofiles==24.1.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.8
configparser==7.1.0
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
frontend==0.0.3
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
isodate==0.6.1
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jmespath==1.0.1
kombu==5.4.0
looseversion==1.3.0
lxml==5.3.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pathlib==1.0.1
pika==1.3.2
playwright==1.46.0
prompt_toolkit==3.0.47
Protego==0.3.1
prov==2.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pydot==3.0.1
pydub==0.25.1
pyee==11.1.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyOpenSSL==24.2.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
queuelib==1.7.0
rdflib==6.3.2
redis==5.0.8
requests==2.31.0
requests-file==2.1.0
scipy==1.13.1
Scrapy==2.11.2
selenium==4.8.3
selenium-recaptcha-solver==1.9.0
service-identity==24.1.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
starlette==0.37.2
tldextract==5.1.2
tomli==2.0.1
traits==6.3.2
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
uvicorn==0.30.6
vine==5.1.0
w3lib==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.0.1
flower
